MenuSider, MenuDrawer 모듈화
